requirements.txt
hydra-core
matplotlib
moviepy
networkx
numpy
opencv-python
Pillow
pyyaml
scienceplots
tensorboard
torchinfo
wandb
